LayoutTests/platform/ios-wk1/http/tests/dom/window-open-about-uppercase-blank-and-access-document-expected.txt

PASS newWindow.document is defined.
